backend/requirements.txt
Flask
Flask-CORS
python-dotenv
bcrypt
PyJWT
SQLAlchemy
psycopg2-binary
alembic
cryptography
